Fix Cout for Fat32 print

+++ FAT32.h
@@ -13,8 +13,8 @@ private:
 	int _SC = 0; // Offset 0D - số byte 1 - Số sector trên cluster 
 	int _SB = 0; // Offset 0E - số byte 2 - Số sector thuộc vừng Bootsector 
 	int _NF = 0; // Offset 10 - số byte 1 - Số bảng FAT, thường là 2 
-	int _SV = 0; // Offset 20 - số byte 4 - Kích thước volume
-	int _SF = 0; // Offset 24 - số byte 4 - Kích thước mỗi bảng FAT 
+	long int _SV = 0; // Offset 20 - số byte 4 - Kích thước volume
+	long int _SF = 0; // Offset 24 - số byte 4 - Kích thước mỗi bảng FAT 
 	int _SCOR = 0; // Offset 2C - số byte 4 - Cluster bắt đầu của RDET
 	int _ExtraInforSector = 0; // Offset 30 - số byte 2 - Sector chứa thông tin phụ (về cluster trống), thường là 1
 	int _BackupBootSector = 0; // Offset 32 - số byte 2 - Sector chứa bản sao lưu của Boot Sector
